+++ TestProjectWeb/Controllers/HomeController.cs
@@ -17,10 +17,5 @@ namespace TestProjectWeb.Controllers
         {
             return View();
         }
-
-        public IActionResult Test() 
-        { 
-            return View(DateTime.Now);
-        }
     }
 }
