+++ TestProjectWeb/Views/Account/AllAccounts.cshtml
@@ -10,7 +10,7 @@
         {
             <div class="user-index-list-item"> 
                 <span class="user-index-list-item-name"> User: @user.Name </span>
-                <a class="user-index-list-item-link" asp-area="" asp-controller="User" asp-action="DeleteUser" asp-route-id="@user.Id">Delete</a>
+                <a class="user-index-list-item-link" asp-area="" asp-controller="Account" asp-action="DeleteAccount" asp-route-id="@user.Id">Delete</a>
             </div>
         }   
     </div>
