+++ TestProjectWeb/Views/Account/Register.cshtml
@@ -2,7 +2,7 @@
 
 <div class="Register">
     Register
-     <form method="post" asp-controller="User" asp-action="Register" asp-antiforgery="true">
+     <form method="post" asp-controller="Account" asp-action="Register" asp-antiforgery="true">
         <div>
             <label asp-for="Name"></label>
             <input type="text" asp-for="Name" />
