--- TestProjectWeb/Views/Home/Test.cshtml
@@ -1,5 +1,0 @@
-﻿@model DateTime
-
-<div>
-    Current date: @Model
-</div>
