+++ TestProjectWeb/Views/Profile/Profile.cshtml
@@ -28,5 +28,5 @@
 </div>
 
 <div>
-    <a asp-area="" asp-controller="User" asp-action="Logout">Logout</a>
+    <a asp-area="" asp-controller="Account" asp-action="Logout">Logout</a>
 </div>
