--- TestProjectWeb/wwwroot/css/User.css
@@ -1,23 +1,0 @@
-.user-index-container {
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-
-    border: 3px solid #000000;
-}
-.user-index-title{
-    font-size: 30px;
-}
-.user-index-list{
-    margin: 5px 0;
-}
-.user-index-list-item{
-
-}
-.user-index-list-item-name{
-    color: #0094ff;
-}
-.user-index-list-item-link{
-    color: #ff0000;
-}
